+++ src/lib.rs
@@ -29,6 +29,8 @@
 //!     Ok(())
 //! }
 //! ```
+#![allow(clippy::return_self_not_must_use)]
+
 mod client;
 pub mod cmd;
 pub mod codec;